Separte google logout button.
